host-macros: Implement a helper proc macro for defining UUIDs cleaner
