+++ host-macros/src/service.rs
@@ -4,59 +4,20 @@
 //! The struct definition is used to define the characteristics of the service, and the ServiceBuilder is used to
 //! generate the code required to create the service.
 
-use darling::{Error, FromMeta};
 use convert_case::{Case, Casing};
+use darling::Error;
 use proc_macro2::TokenStream as TokenStream2;
 use quote::{format_ident, quote, quote_spanned, ToTokens};
 use syn::parse::Result;
 use syn::spanned::Spanned;
-use syn::{Expr, Meta, Token};
+use syn::{Meta, Token};
 
 use crate::characteristic::{AccessArgs, Characteristic};
-use crate::uuid::Uuid;
+use crate::uuid::parse_arg_uuid;
 
 #[derive(Debug)]
 pub(crate) struct ServiceArgs {
     pub uuid: TokenStream2,
-}
-
-/// Parse the UUID argument of the service attribute.
-///
-/// The UUID can be specified as a string literal, an integer literal, or an expression that impl Into<Uuid>.
-fn parse_arg_uuid(value: &Expr) -> Result<TokenStream2> {
-    match value {
-        Expr::Lit(lit) => {
-            if let syn::Lit::Str(lit_str) = &lit.lit {
-                let uuid_string = Uuid::from_string(&lit_str.value()).map_err(|_| {
-                    Error::custom(
-                        "Invalid UUID string.  Expect i.e. \"180f\" or \"0000180f-0000-1000-8000-00805f9b34fb\"",
-                    )
-                    .with_span(&lit.span())
-                })?;
-                Ok(quote::quote! {#uuid_string})
-            } else if let syn::Lit::Int(lit_int) = &lit.lit {
-                let uuid_string = Uuid::Uuid16(lit_int.base10_parse::<u16>().map_err(|_| {
-                    Error::custom("Invalid 16bit UUID literal.  Expect i.e. \"0x180f\"").with_span(&lit.span())
-                })?);
-                Ok(quote::quote! {#uuid_string})
-            } else {
-                Err(Error::custom(
-                    "Invalid UUID literal.  Expect i.e. \"180f\" or \"0000180f-0000-1000-8000-00805f9b34fb\"",
-                )
-                .with_span(&lit.span())
-                .into())
-            }
-        }
-        other => {
-            let span = other.span(); // span will highlight if the value does not impl Into<Uuid>
-            Ok(quote::quote_spanned! { span =>
-                {
-                    let uuid: trouble_host::types::uuid::Uuid = #other.into();
-                    uuid
-                }
-            })
-        }
-    }
 }
 
 impl syn::parse::Parse for ServiceArgs {
